switch other parts of job-view to use push utility function
